requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
playwright==1.40.0
pydantic==2.5.0
python-dotenv==1.0.0
python-dateutil==2.8.2
Pillow==10.0.1